refactor: update service imports to use specific service modules for user, payment, and workout functionalities
