+++ src/app/hooks/useSubscription.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { fetchUserSubscription } from "@/services/api";
+import { fetchUserSubscription } from "@/services/paymentService";
 
 export const useSubscription = (enabled: boolean) => {
   const { data, isLoading } = useQuery({
