+++ src/services/userService.ts
@@ -1,0 +1,22 @@
+const BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
+
+export async function fetchUserDetails() {
+  try {
+    const response = await fetch(`${BASE_URL}/api/user-info`, {
+      method: 'GET',
+      credentials: 'include',
+    });
+    if (response.status === 401) {
+      throw new Error('Unauthorized: Please log in');
+    }
+    if (!response.ok) {
+      throw new Error('Failed to fetch user details');
+    }
+
+    const data = await response.json();
+    return data;
+  } catch (error) {
+    console.error('Error fetching user details:', error);
+    throw error;
+  }
+}
